Add check for whether user account is active
I previously added a new field named 'active' to the user model. In case this field is set to false, a user is not able to login as his account is suspended.

To implement this, this has to override the laravel login() function in the LoginController.
Also suitable error messages are added to be displayed in the login view (auth.login) in case a disabled account tries to log in.