Improve quality of Authentication middleware
The authentication middelware now is more dynamic. The IsAdmin, IsTeacher and IsStudent middleware classes are removed and replaced by a HasRole middelware accessible with 'role:[role]'.
[role] can be any string, but the role check is only successful if the checked user object has set the given role.

I tried adding a middleware group for the auth middleware and the new role middleware. This did not work due to the [role] parameter needed for the role middleware; I could not get the middleware group to accept a parameter.
To avoid calling the auth middleware every time a role is checked, I also check if a user session is active within the HasRole middleware.

If the HasRole middleware fails, the page is aborted with an 404 error.

I also changed a comment in the User.php file (code quality).